added NIIRS column to RasterEntryMetadata.    It is a nullable field and is of type Double
git-svn-id: http://svn.osgeo.org/ossim/omar@16874 b630af17-092b-0410-afba-a002956477a4

+++ apps/omar/src/groovy/org/ossim/omar/OmsInfoParser.groovy
@@ -325,7 +325,7 @@ public class OmsInfoParser
               rasterEntry.metadata.description = value
               break;
             case "niirs":
-              rasterEntry.metadata.niirs = value
+              rasterEntry.metadata.niirs = value as Double
               break;
 
           // Just for testing
